fix/절취선 제거,출력클래스 OutputView에서 분리=>RoomCheckOutput
